Move over the square diagonal example
